+++ src/server/update.rs
@@ -1,0 +1,193 @@
+use log::{debug, error, info};
+
+use std::{fs, io::ErrorKind, path::Path, process::exit};
+
+use crate::{
+  constants, files::config::load_config, server, steamcmd::steamcmd_command, utils::get_working_dir,
+};
+
+#[derive(Clone, Debug)]
+pub struct UpdateInfo {
+  current_buildid: String,
+  latest_buildid: String,
+}
+
+impl UpdateInfo {
+  pub fn new() -> Self {
+    let current_buildid = get_current_buildid();
+    let latest_buildid = get_latest_buildid();
+
+    Self {
+      current_buildid,
+      latest_buildid,
+    }
+  }
+
+  pub fn update_available(&self) -> bool {
+    self.current_buildid != self.latest_buildid
+  }
+
+  // pub fn current_buildid(&self) -> String {
+  //   self.current_buildid.clone()
+  // }
+
+  // pub fn latest_buildid(&self) -> String {
+  //   self.latest_buildid.clone()
+  // }
+}
+
+pub fn update_is_available() -> bool {
+  let info = UpdateInfo::new();
+  debug!("{:#?}", info);
+
+  info.update_available()
+}
+
+pub fn update_server() {
+  // Shutdown the server if it's running
+  let server_was_running = server::is_running();
+  if server_was_running {
+    server::blocking_shutdown();
+  }
+
+  // Update the installation
+  if let Err(e) = server::install(constants::GAME_ID) {
+    error!("Failed to install server: {}", e);
+    exit(1);
+  }
+
+  // Bring the server up if it was running before
+  if server_was_running {
+    let config = load_config();
+    match server::start_daemonized(config) {
+      Ok(_) => info!("Server daemon started"),
+      Err(e) => {
+        error!("Error daemonizing: {}", e);
+        exit(1);
+      }
+    }
+  }
+}
+
+fn get_current_buildid() -> String {
+  let manifest_path = Path::new(&get_working_dir())
+    .join("steamapps")
+    .join(&format!("appmanifest_{}.acf", constants::GAME_ID));
+  let manifest_data = fs::read_to_string(&manifest_path).unwrap_or_else(|_| {
+    panic!(
+      "Failed to read manifest file at '{}'",
+      manifest_path.display()
+    )
+  });
+  extract_buildid_from_manifest(&manifest_data).to_string()
+}
+
+fn get_latest_buildid() -> String {
+  // Remove the cached file to force an updated response. This is done because `steamcmd` seems to
+  // refuse to update information before querying the app_info even with `+app_info_update 1` or
+  // `+@bCSForceNoCache 1`
+  let appinfo_file = Path::new("/home/steam/Steam/appcache/appinfo.vdf");
+  fs::remove_file(&appinfo_file).unwrap_or_else(|e| match e.kind() {
+    // AOK if it doesn't exist
+    ErrorKind::NotFound => {}
+    err_kind => {
+      error!(
+        "Failed to remove appinfo file at '{}'! Error: {:?}",
+        appinfo_file.display(),
+        err_kind
+      );
+      exit(1);
+    }
+  });
+
+  // Now pull the latest app info
+  let args = &[
+    "+@ShutdownOnFailedCommand 1",
+    "+login anonymous",
+    &format!("+app_info_print {}", constants::GAME_ID),
+    "+quit",
+  ];
+  let mut steamcmd = steamcmd_command();
+  let app_info_output = steamcmd
+    .args(args)
+    .output()
+    .expect("Failed to run steamcmd");
+  assert!(app_info_output.status.success());
+
+  let stdout = String::from_utf8(app_info_output.stdout).expect("steamcmd returned invalid UTF-8");
+  extract_buildid_from_app_info(&stdout).to_string()
+}
+
+fn extract_buildid_from_manifest(manifest: &str) -> &str {
+  for line in manifest.lines() {
+    if line.trim().starts_with("\"buildid\"") {
+      return split_vdf_key_val(line).1;
+    }
+  }
+
+  panic!("Unexpected manifest format:\n{}", manifest);
+}
+
+fn extract_buildid_from_app_info(app_info: &str) -> &str {
+  let mut lines = app_info.lines();
+  while let Some(line) = lines.next() {
+    if line.trim() == "\"public\"" {
+      break;
+    }
+  }
+
+  assert_eq!(
+    lines.next().map(|line| line.trim()),
+    Some("{"),
+    "Invalid app info"
+  );
+  let build_id_line = lines
+    .next()
+    .unwrap_or_else(|| panic!("Invalid app info format:\n{}", app_info))
+    .trim();
+  assert!(build_id_line.starts_with("\"buildid\""), "Invalid app info");
+
+  split_vdf_key_val(build_id_line).1
+}
+
+// Note: This is super brittle and will fail if there is whitespace within the key or value _or_ if
+// there are escaped " at the end of the key or value
+fn split_vdf_key_val(vdf_pair: &str) -> (&str, &str) {
+  let mut pieces = vdf_pair.trim().split_whitespace();
+  let key = pieces.next().expect("Missing vdf key").trim_matches('"');
+  let val = pieces.next().expect("Missing vdf val").trim_matches('"');
+
+  (key, val)
+}
+
+#[cfg(test)]
+mod tests {
+  use super::*;
+
+  use once_cell::sync::Lazy;
+
+  use std::path::PathBuf;
+
+  static TEST_ASSET_DIR: Lazy<PathBuf> = Lazy::new(|| {
+    Path::new(env!("CARGO_MANIFEST_DIR"))
+      .join("tests")
+      .join("assets")
+  });
+
+  #[test]
+  fn extracting_buildid_from_manifest() {
+    let sample_file = TEST_ASSET_DIR.join("example_app_manifest.txt");
+    let manifest_data = fs::read_to_string(sample_file).expect("Sample manifest file missing");
+
+    assert_eq!(extract_buildid_from_manifest(&manifest_data), "6246034");
+  }
+
+  #[test]
+  fn extracting_buildid_from_app_info() {
+    let sample_file = TEST_ASSET_DIR.join("example_steamcmd_app_info.txt");
+    let app_info_output =
+      fs::read_to_string(sample_file).expect("Sample app info output file missing");
+
+    assert_eq!(extract_buildid_from_app_info(&app_info_output), "6246034");
+  }
+}
